# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KISS Notebook API",
    "version": "1.0.0",
    "description": "REST API for collaborative Jupyter-like notebook platform.\n\nAll successful responses are wrapped in `{\"data\": ...}`; errors in `{\"error\": \"message\"}`.\n\n**Auth flow in Swagger UI:**\n1. `POST /api/v1/auth/login` — response sets `session_id` and `csrf_token` cookies.\n2. Copy the `csrf_token` cookie value.\n3. Click **Authorize**: set `cookieAuth` = `session_id` value, set `csrfToken` = `csrf_token` value.\n4. All subsequent mutating requests will include `X-CSRF-Token` header automatically."
  },
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/notebooks": {
      "get": {
        "tags": ["notebooks"],
        "summary": "List user notebooks",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Title substring filter" }
        ],
        "responses": {
          "200": {
            "description": "Paginated notebook list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNotebookListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["notebooks"],
        "summary": "Create a notebook",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateNotebookRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notebook created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNotebookResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/notebooks/shared": {
      "get": {
        "tags": ["notebooks"],
        "summary": "List notebooks shared with the current user",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of shared notebooks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNotebookListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/notebooks/{id}": {
      "get": {
        "tags": ["notebooks"],
        "summary": "Get notebook by ID (with blocks)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "responses": {
          "200": {
            "description": "Notebook with blocks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNotebookResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["notebooks"],
        "summary": "Update notebook",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateNotebookRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notebook updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNotebookResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["notebooks"],
        "summary": "Delete notebook",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "responses": {
          "200": {
            "description": "Notebook deleted (returns `{\"data\": null}`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNullResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/notebooks/{id}/blocks": {
      "post": {
        "tags": ["blocks"],
        "summary": "Add block to notebook",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBlockRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Block created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataBlockResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/notebooks/{id}/blocks/{blockID}": {
      "put": {
        "tags": ["blocks"],
        "summary": "Update block",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/NotebookID" },
          { "$ref": "#/components/parameters/BlockID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateBlockRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Block updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataBlockResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["blocks"],
        "summary": "Delete block",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/NotebookID" },
          { "$ref": "#/components/parameters/BlockID" }
        ],
        "responses": {
          "200": {
            "description": "Block deleted (returns `{\"data\": null}`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNullResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/notebooks/{id}/permissions": {
      "get": {
        "tags": ["permissions"],
        "summary": "List permissions granted on a notebook",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "responses": {
          "200": {
            "description": "Permission list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataPermissionListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/notebooks/{id}/permissions/invite": {
      "post": {
        "tags": ["permissions"],
        "summary": "Grant permission to a user by email or username",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GrantPermissionByIdentifierRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission granted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataPermissionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/notebooks/{id}/permissions/{userID}": {
      "put": {
        "tags": ["permissions"],
        "summary": "Grant a permission level to a user on a notebook",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/NotebookID" },
          { "$ref": "#/components/parameters/UserID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GrantPermissionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission granted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataPermissionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["permissions"],
        "summary": "Revoke a user's permission on a notebook",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/NotebookID" },
          { "$ref": "#/components/parameters/UserID" }
        ],
        "responses": {
          "204": { "description": "Permission revoked" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    },
    "parameters": {
      "NotebookID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
      "BlockID": {
        "name": "blockID",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
      "UserID": {
        "name": "userID",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DataNullResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "null", "nullable": true }
        }
      },
...
      "CreateNotebookRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" }
        }
      },
      "UpdateNotebookRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "maxLength": 255 },
          "is_public": { "type": "boolean" }
        }
      },
      "NotebookResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "owner_id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "is_public": { "type": "boolean" },
          "blocks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BlockResponse" }
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "DataNotebookResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/NotebookResponse" }
        }
      },
      "NotebookListResponse": {
        "type": "object",
        "properties": {
          "notebooks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NotebookResponse" }
          },
          "total": { "type": "integer" },
          "limit": { "type": "integer" },
          "offset": { "type": "integer" }
        }
      },
      "DataNotebookListResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/NotebookListResponse" }
        }
      },
      "CreateBlockRequest": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "description": "e.g. `code`, `text`" },
          "language": { "type": "string", "description": "e.g. `python`, `markdown`" },
          "content": { "type": "string" }
        }
      },
      "UpdateBlockRequest": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "language": { "type": "string" },
          "content": { "type": "string" }
        }
      },
      "BlockResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "type": { "type": "string" },
          "language": { "type": "string" },
          "content": { "type": "string" },
          "position": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "DataBlockResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/BlockResponse" }
        }
      },
      "GrantPermissionRequest": {
        "type": "object",
        "required": ["level"],
        "properties": {
          "level": {
            "type": "string",
            "enum": ["read", "write"],
            "description": "Permission level to grant"
          }
        }
      },
      "GrantPermissionByIdentifierRequest": {
        "type": "object",
        "required": ["identifier", "level"],
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email or username of the user to grant access to"
          },
          "level": {
            "type": "string",
            "enum": ["readonly", "editor"],
            "description": "Permission level to grant"
          }
        }
      },
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "notebook_id": { "type": "integer", "format": "int64" },
          "user_id": { "type": "integer", "format": "int64" },
          "permission_level": { "type": "string", "enum": ["read", "write"] }
        }
      },
      "DataPermissionResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PermissionResponse" }
        }
      },
      "PermissionListResponse": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PermissionResponse" }
          }
        }
      },
      "DataPermissionListResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PermissionListResponse" }
        }
      },
...
    }
  }
}